Extract v3 migration skill installation
Move temporary v3 migration skill detection, install, removal, and specific-skill helpers into V3MigrationSkillInstaller. Keep ToolSkillSynchronizer as the service-facing facade for migration operations.
