+++ Packages/src/Editor/Infrastructure/SkillSetup/V3MigrationSkillInstaller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1fdf210a2fe94c1ca9d8d122c895e57
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
